main screen fills with db data
added a constructer to IouItem to take in an Iou. IouItem stores a copy
of the actual iou so it can be edited/deleted and the data from the db
can be adjusted accordingly. no access functions to it yet, but was not
sure how other activities would handle it